refactor : move BaseBallGame directory location
